:truck: [ refactor ] 상수파일 카멜케이스로 변경
refacotr
